tests: cleanup bogus tests
