+++ packages/core/tests/plugin/plugin-entry.test.ts
@@ -4,10 +4,6 @@ import * as pluginEntry from '#src/plugin.index';
 
 // the phase enums' single home is the core root barrel
 describe('@seedcord/core/plugin entry', () => {
-    it('exports the Plugin base', () => {
-        expect(pluginEntry.Plugin).toBeTypeOf('function');
-    });
-
     it('keeps internal helpers and the phase enums off the entry', () => {
         expect(Object.keys(pluginEntry)).toStrictEqual(['Plugin']);
     });
